Untangle the mess around options/command/screens a little
There was a bit of a discussion on doku-intern that the current
wording/examples on this topic were unclear. To be fair, the example
was written during the DocBook 4 era when <option/> within <command/>
was still allowed. But now it is just confusing.

The rule around in-text command+option combinations had been changed
post-DocBook-5 introduction but was always a bit clumsy, so that is
now simplified too.