Structure section: reformat DocBook examples to 1-space indent
We somehow had a 2-space indent for the examples, which on its own
makes sense (and I'd actually wish we'd use it) but is against the
rules laid out in the XML formatting chapter.
Also fixes a few other minor issues with examples, like a missing end
tag for the "example" example.