Lists: Align list terminology with DocBook tags
The previous list terminology was a bit weird in that it was different
from Docbook every step of the way and was also sometimes inexact, e.g.
not all ordered lists are actually numbered, lower levels will have
alphabetical characters etc.

We now use the same terminology as DocBook, so at least the tag names
are the same as the descriptive names, even though that is also not
always ideal, e.g. itemized list v/ unordered list is a weird one, as
is descriptive v/ variable list.